Refactor `this` magic out of testing jsx function
